Moved Storage exceptions to storage/exceptions package
