+++ src/main/java/uk/ac/ebi/ampt2d/storage/StorageService.java
@@ -2,6 +2,8 @@ package uk.ac.ebi.ampt2d.storage;
 
 import org.springframework.core.io.Resource;
 import org.springframework.web.multipart.MultipartFile;
+import uk.ac.ebi.ampt2d.storage.exceptions.StorageException;
+import uk.ac.ebi.ampt2d.storage.exceptions.StorageFileNotFoundException;
 
 import java.nio.file.Path;
