+++ src/main/java/uk/ac/ebi/ampt2d/storage/exceptions/StorageFileNotFoundException.java
@@ -1,4 +1,6 @@
-package uk.ac.ebi.ampt2d.storage;
+package uk.ac.ebi.ampt2d.storage.exceptions;
+
+import uk.ac.ebi.ampt2d.storage.exceptions.StorageException;
 
 public class StorageFileNotFoundException extends StorageException {
